requirements.txt
streamlit
ultralytics
Pillow
numpy
opencv-python-headless
torch --index-url https://download.pytorch.org/whl/cpu
torchvision
gdown